[mypyc] Infer native int index type from range over native int (#13640)
Previously the for index variable would always be inferred a plain `int`
when iterating over range. Now we infer a native int type if one of the
range arguments is a native int.

Example:
```
for x in range(i64(5)):
    # type of x is i64
    ...
```

This also works in comprehensions. This is particularly useful since
there's no way to annotate the index variables in comprehensions, as
they are in an inner scope.

This is a convenience feature that makes it easier to avoid implicit
conversions between int and native int types.

Work on mypyc/mypyc#837.