Tweak constraints handling for splitting typevartuples (#13716)
The existing logic for splitting mapped & template into prefix, middle,
and suffix does not handle the case where the template middle itself is
not a singleton unpack but rather itself has a prefix & suffix.

In this case we need to pull out the prefix & suffix by doing a second
round of splitting on the middle. Originally we weren't sure if the PEP
required implementing this double split, but one of the PEP646 test
cases requires it.