Silence errors from third party packages in daemon (#13768)
If daemon was running and additional third party code was installed,
mypy daemon would report errors from the code. This was inconsistent 
with non-daemon mypy and the initial daemon run. If mypy was run with 
strict options, this could results lots of errors being generated.

The same code is used to silence errors from typeshed. I'm using
typeshed in the test case since it was easier to do.

I also manually verified that this works with pytest, aiohttp and some
typeshed stub packages.

Fixes #13140.