Handle attrs' __attrs_init__ method (#13865)
If an `attrs` class does not generate the `__init__` method for whatever
reason, the method is still actually generated, under the name
`__attrs_init__`. This is intended to allow a user `__init__` to then
delegate to the `attrs` implementation to do the assignment (especially
useful with frozen classes). This PR makes Mypy typecheck this method in
this situation.