Implement basic *args support for variadic generics (#13889)
This implements the most basic support for the *args feature but various
edge cases are not handled in this PR because of the large volume of
places that needed to be modified to support this.

In particular, we need to special handle the ARG_STAR argument in
several places for the case where the type is a UnpackType. Finally when
we actually check a function we need to construct a TupleType instead of
a builtins.tuple.