[mypyc] Fix glue methods with native int types (#13939)
This fixes handling of bitmap arguments used to track default values of
arguments with native int types in method overrides.

The main use case is a method override that adds a native int argument
with a default value to the base class signature.

Also generate an error if we don't support generating a glue method for
an override. This can happen if the positions of bits in the bitmap
would change in the subclass. We also don't support switching between
error values and default value bitmaps in glue methods. These could be
supported, but it would be complicated and doesn't seem worth it.

Work on mypyc/mypyc#837.