Fix new style union syntax in type aliases (#14008)
Fix Python 3.10 `|` union syntax in type aliases, when one of
the operands is a type alias or a type with an overloaded `__init__`.

We can now infer `typing._SpecialForm` for type aliases in a runtime 
context.

Also create a bunch of minimal test-only stubs for stdlib modules 
to fix some test failures caused by the missing `typing._SpecialForm`
in the default test stubs. This is generally what we want in any case, 
since using typeshed stubs with minimal builtins/typing stubs can
result in unpredictable behavior and slow tests.

Fixes #12368. Fixes #12005. Fixes #11426.