Try empty context when assigning to union typed variables (#14151)
Fixes #4805
Fixes #13936

It is known that mypy can overuse outer type context sometimes
(especially when it is a union). This prevents a common use case for
narrowing types (see issue and test cases). This is a somewhat major
semantic change, but I think it should match better what a user would
expect.