Switch error code used to report vars defined in different branch (#14176)
We previously used `use-before-def` code here but this commit switched
it to use `partially-defined`. This particular check generates a lot of
false positives, in particular around loops of the form:
```python
for i in range(2)
  if i == 0:
    x = 1
  else:
    y = x
```

While in an ideal world mypy has no false positives, it's not feasible
for us to handle this correctly in the short-term. Moving this to
partially-defined error code makes the `use-before-def` have a much
lower false positive rate, which is a plus. Unfortunately,
`partially-defined` will always have a higher false positive rate. This
means that if we enable it by default, lots of people will disable this
check. We want to avoid the same thing happening to use-before-def
check.

See [this
PR](https://github.com/python/mypy/pull/14166#issuecomment-1325709734)
for further discussion.