Use secure HTTPS link for Mypy badge (#14383)
Now that the site is accessible via HTTPS

 - #9919

Fetch the badge using a secure link, especially since this will be used
an example for others. Also, non-secure badges will not be shown in some
contexts, such as DockerHub, so this makes the example both more secure
and more generally useful.