Improve CallableType handling with typevar tuples (#14465)
Mainly adds some tests that were previously disabled since they failed,
as well as tests for handling prefices. Suffices tests are not added yet
because when a suffix is involved we can't use the existing solution
since it will construct positional args after *args, when we really need
something like `*tuple[*args, <suffix>]` in order to encode this in the
type info.

We also refactor out the manipulation of the arg names/kinds/types into
its own helper which will make control flow less annoying to deal with.

We are deferring the implementation of suffices until more parts of the
PEP are implemented since it doesn't seem as important (e.g. in other
parts of the language this is not supported)