[dataclass_transform] support field_specifiers (#14667)
These are analogous to `dataclasses.field`/`dataclasses.Field`.

Like most dataclass_transform features so far, this commit mostly just
plumbs through the necessary metadata so that we can re-use the existing
`dataclasses` plugin logic. It also adds support for the `alias=` and
`factory=` kwargs for fields, which are small; we rely on typeshed to
enforce that these aren't used with `dataclasses.field`.