Reduce size of JSON data in Mypy cache (#14808)
By default, Python outputs spaces after commas and colons when dumping
JSON data. This is normally fine, but when exporting large amounts of
JSON (such as cache info) it is more space efficient to export JSON
without whitespaces. To do this, simply pass `separators=(",", ":")` to
`json.dumps()`.

Although the space savings aren't massive, they do reduce the size of
the cache folder by a couple of megabytes:

| Method  | Size   |
|---------|--------|
| Current | 31.2MB |
| This PR | 29.3MB |

For comparison, the size of the cache folder with the `--sqlite-cache`
option is about 27.5MB.